+++ src/parser/parser.rs
@@ -524,7 +524,7 @@ impl<'a> SysMLParser for Parser<'a> {
     }
 
     fn parse_identification(&mut self) {
-        super::grammar::kerml::parse_identification(self)
+        super::grammar::sysml::parse_identification(self)
     }
 
     fn parse_body(&mut self) {